% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelStMoE.R
\docType{class}
\name{ModelStMoE-class}
\alias{ModelStMoE-class}
\alias{ModelStMoE}
\title{A Reference Class which represents a fitted StMoE model.}
\description{
ModelMRHLP represents a \link[=ModelStMoE]{StMoE} model for which parameters have
been estimated.
}
\section{Fields}{

\describe{
\item{\code{paramMRHLP}}{A \link{ParamStMoE} object. It contains the estimated values of the parameters.}

\item{\code{statMRHLP}}{A \link{StatStMoE} object. It contains all the statistics associated to the StMoE model.}
}}


\seealso{
\link{ParamStMoE}, \link{StatStMoE}
}
